CIFF parsing works

+++ parser.cpp
@@ -18,7 +18,7 @@ std::string Parser::read_magic(std::vector<unsigned char> buffer){
 }
 
 std::vector<unsigned char>* Parser::read_header(std::vector<unsigned char> buffer,int from,int count){
-    if((from+count) >= buffer.size()){
+    if((long unsigned int)(from+count) >= buffer.size()){
         throw new std::out_of_range("Too long read!");
     }
     std::vector<unsigned char>* header_buffer = new std::vector<unsigned char>;
